+++ db/updates/update_12.4-12.5.sql
@@ -56,5 +56,9 @@ UPDATE `suppliers` SET `owner_id` = '1';
 UPDATE `synonyms` SET `owner_id` = '1'; 
 UPDATE `templates` SET `owner_id` = '1'; 
 
+ALTER TABLE `ingredient_safety_data` DROP PRIMARY KEY;
+ALTER TABLE `ingredient_safety_data` ADD `id` INT NOT NULL AUTO_INCREMENT FIRST, ADD PRIMARY KEY (`id`); 
+ALTER TABLE `ingredient_safety_data` ADD `updated_at` TIMESTAMP on update CURRENT_TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP AFTER `created_at`; 
+
 ALTER TABLE `bottles` CHANGE `ml` `ml` DOUBLE NOT NULL; 
 ALTER TABLE `bottles` CHANGE `height` `height` DOUBLE NULL DEFAULT 0;