+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Improve PubChem structure images fetch
 - Fixed redirection when adding ingredient if already exists with the same name
 - Update ingredient duplication function
+- Change default primary key for safety data
 
 ### Version 12.4
 - Add dilutant in API formulas